+++ Scripts/HumanInterface/HISpidarBehaviour.cs
@@ -81,7 +81,7 @@ public class HISpidarBehaviour : SprBehaviour {
 
                 if (lengthText != null) {
                     string text = "";
-                    for (int i = 0; i < hiSpidar.NMotor(); i++) {
+                    for (int i = 0; i < (int)hiSpidar.NMotor(); i++) {
                         text += hiSpidar.GetMotor((uint)i).GetLength().ToString() + "\r\n";
                     }
                     lengthText.text = text;
